language_model: Invert dep on client
